# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Memozapia API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Note {
  id: number;
  title?: string | null;
  content: string;
  tags: string[];
  scheduled_at?: string | null;
  gcal_event_id?: string | null;
  created_at: string;
  updated_at: string;
}

export interface Tag {
  id: number;
  name: string;
}

export interface CreateNoteBody {
  title?: string | null;
  content: string;
  tags?: string[];
  scheduled_at?: string | null;
}

export interface UpdateNoteBody {
  title?: string | null;
  content: string;
  tags?: string[];
  scheduled_at?: string | null;
}

export type ListNotesParams = {
  search?: string;
  tag?: string;
};

export type DeleteNote200 = {
  message: string;
};