# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Memozapia API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: notes
    description: Note operations
  - name: tags
    description: Tag operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes
      description: Returns notes, optionally filtered by search term or tag
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  /notes/tags/all:
    get:
      operationId: listTags
      tags: [tags]
      summary: Get all unique tags
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"

  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
    put:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteBody"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          nullable: true
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        scheduled_at:
          type: string
          nullable: true
        gcal_event_id:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - content
        - tags
        - created_at
        - updated_at

    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name

    CreateNoteBody:
      type: object
      properties:
        title:
          type: string
          nullable: true
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        scheduled_at:
          type: string
          nullable: true
      required:
        - content

    UpdateNoteBody:
      type: object
      properties:
        title:
          type: string
          nullable: true
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        scheduled_at:
          type: string
          nullable: true
      required:
        - content