# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Memozapia API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns notes, optionally filtered by search term or tag
 * @summary List notes
 */
export const ListNotesQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  tag: zod.coerce.string().optional(),
});

export const ListNotesResponseItem = zod.object({
  id: zod.number(),
  title: zod.string().nullish(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  scheduled_at: zod.string().nullish(),
  gcal_event_id: zod.string().nullish(),
  created_at: zod.string(),
  updated_at: zod.string(),
});
export const ListNotesResponse = zod.array(ListNotesResponseItem);

/**
 * @summary Create a note
 */
export const CreateNoteBody = zod.object({
  title: zod.string().nullish(),
  content: zod.string(),
  tags: zod.array(zod.string()).optional(),
  scheduled_at: zod.string().nullish(),
});

/**
 * @summary Get all unique tags
 */
export const ListTagsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
});
export const ListTagsResponse = zod.array(ListTagsResponseItem);

/**
 * @summary Get a note by ID
 */
export const GetNoteParams = zod.object({
  id: zod.coerce.number(),
});

export const GetNoteResponse = zod.object({
  id: zod.number(),
  title: zod.string().nullish(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  scheduled_at: zod.string().nullish(),
  gcal_event_id: zod.string().nullish(),
  created_at: zod.string(),
  updated_at: zod.string(),
});

/**
 * @summary Update a note
 */
export const UpdateNoteParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateNoteBody = zod.object({
  title: zod.string().nullish(),
  content: zod.string(),
  tags: zod.array(zod.string()).optional(),
  scheduled_at: zod.string().nullish(),
});

export const UpdateNoteResponse = zod.object({
  id: zod.number(),
  title: zod.string().nullish(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  scheduled_at: zod.string().nullish(),
  gcal_event_id: zod.string().nullish(),
  created_at: zod.string(),
  updated_at: zod.string(),
});

/**
 * @summary Delete a note
 */
export const DeleteNoteParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteNoteResponse = zod.object({
  message: zod.string(),
});