# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Memozapia API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: number;
  title?: string | null;
  content: string;
  tags: string[];
  scheduled_at?: string | null;
  gcal_event_id?: string | null;
  created_at: string;
  updated_at: string;
}